requirements.txt
django-filter
Django==2.2.7
django-tables2==2.2.0
pytz==2019.3
sqlparse==0.3.0
